requirements.txt
streamlit
python-dotenv
openai
groq